Save one executeStatementStep for very simple statement
