+++ herddb-core/src/main/java/herddb/model/planner/BindableTableScanOp.java
@@ -19,54 +19,17 @@
  */
 package herddb.model.planner;
 
-import herddb.core.TableSpaceManager;
-import herddb.model.DataScanner;
-import herddb.model.Projection;
-import herddb.model.ScanResult;
-import herddb.model.StatementEvaluationContext;
-import herddb.model.StatementExecutionException;
-import herddb.model.StatementExecutionResult;
-import herddb.model.TransactionContext;
 import herddb.model.commands.ScanStatement;
-import herddb.utils.Wrapper;
 
 /**
  * TableScanOp + ProjectOp + FilerOp
  *
  * @author eolivelli
  */
-public class BindableTableScanOp implements PlannerOp {
-
-    final ScanStatement statement;
+public class BindableTableScanOp extends SimpleScanOp {
 
     public BindableTableScanOp(ScanStatement scan) {
-        this.statement = scan;
-    }
-
-    @Override
-    public String getTablespace() {
-        return statement.getTableSpace();
-    }
-
-    public ScanStatement getStatement() {
-        return statement;
-    }
-
-    @Override
-    public StatementExecutionResult execute(TableSpaceManager tableSpaceManager,
-        TransactionContext transactionContext, StatementEvaluationContext context,
-        boolean lockRequired, boolean forWrite) throws StatementExecutionException {
-        DataScanner scan = tableSpaceManager.scan(statement, context, transactionContext, lockRequired, forWrite);
-        return new ScanResult(transactionContext.transactionId, scan);
-    }
-
-    @Override
-    public <T> T unwrap(Class<T> clazz) {
-        T unwrapped = statement.unwrap(clazz);
-        if (unwrapped != null) {
-            return unwrapped;
-        }
-        return Wrapper.unwrap(this, clazz);
+        super(scan);
     }
 
     @Override
