+++ herddb-core/src/main/java/herddb/model/planner/FilteredTableScanOp.java
@@ -62,7 +62,7 @@ public class FilteredTableScanOp implements PlannerOp {
             TransactionContext transactionContext, StatementEvaluationContext context,
             boolean lockRequired, boolean forWrite) throws StatementExecutionException {
         DataScanner scan = tableSpaceManager.scan(statement, context, transactionContext,
-                 lockRequired, forWrite);
+                lockRequired, forWrite);
         return new ScanResult(transactionContext.transactionId, scan);
     }
 
@@ -73,5 +73,10 @@ public class FilteredTableScanOp implements PlannerOp {
             return unwrapped;
         }
         return Wrapper.unwrap(this, clazz);
+    }
+
+    @Override
+    public boolean isSimpleStatementWrapper() {
+        return true;
     }
 }
