+++ herddb-core/src/main/java/herddb/model/planner/LimitedBindableTableScanOp.java
@@ -35,38 +35,10 @@ import herddb.utils.Wrapper;
  *
  * @author eolivelli
  */
-public class LimitedBindableTableScanOp implements PlannerOp {
-
-    final ScanStatement statement;
+public class LimitedBindableTableScanOp extends SimpleScanOp {
 
     public LimitedBindableTableScanOp(ScanStatement scan) {
-        this.statement = scan;
-    }
-
-    @Override
-    public String getTablespace() {
-        return statement.getTableSpace();
-    }
-
-    public ScanStatement getStatement() {
-        return statement;
-    }
-
-    @Override
-    public StatementExecutionResult execute(TableSpaceManager tableSpaceManager,
-        TransactionContext transactionContext,
-        StatementEvaluationContext context, boolean lockRequired, boolean forWrite) throws StatementExecutionException {
-        DataScanner scan = tableSpaceManager.scan(statement, context, transactionContext, lockRequired, forWrite);
-        return new ScanResult(transactionContext.transactionId, scan);
-    }
-
-    @Override
-    public <T> T unwrap(Class<T> clazz) {
-        T unwrapped = statement.unwrap(clazz);
-        if (unwrapped != null) {
-            return unwrapped;
-        }
-        return Wrapper.unwrap(this, clazz);
+        super(scan);
     }
 
     @Override
