+++ herddb-core/src/main/java/herddb/model/planner/ProjectedTableScanOp.java
@@ -53,8 +53,8 @@ public class ProjectedTableScanOp implements PlannerOp {
 
     @Override
     public StatementExecutionResult execute(TableSpaceManager tableSpaceManager,
-        TransactionContext transactionContext,
-        StatementEvaluationContext context, boolean lockRequired, boolean forWrite) throws StatementExecutionException {
+            TransactionContext transactionContext,
+            StatementEvaluationContext context, boolean lockRequired, boolean forWrite) throws StatementExecutionException {
         DataScanner scan = tableSpaceManager.scan(statement, context, transactionContext, lockRequired, forWrite);
         return new ScanResult(transactionContext.transactionId, scan);
     }
@@ -66,6 +66,11 @@ public class ProjectedTableScanOp implements PlannerOp {
             return unwrapped;
         }
         return Wrapper.unwrap(this, clazz);
+    }
+
+    @Override
+    public boolean isSimpleStatementWrapper() {
+        return true;
     }
 
     @Override
