+++ herddb-core/src/main/java/herddb/model/planner/SimpleDeleteOp.java
@@ -62,4 +62,10 @@ public class SimpleDeleteOp implements PlannerOp {
         }
         return Wrapper.unwrap(this, clazz);
     }
+
+    @Override
+    public boolean isSimpleStatementWrapper() {
+        return true;
+    }
+
 }
