+++ herddb-core/src/main/java/herddb/model/planner/SortedBindableTableScanOp.java
@@ -21,7 +21,6 @@ package herddb.model.planner;
 
 import herddb.core.TableSpaceManager;
 import herddb.model.DataScanner;
-import herddb.model.Projection;
 import herddb.model.ScanResult;
 import herddb.model.StatementEvaluationContext;
 import herddb.model.StatementExecutionException;
@@ -35,38 +34,10 @@ import herddb.utils.Wrapper;
  *
  * @author eolivelli
  */
-public class SortedBindableTableScanOp implements PlannerOp {
-
-    final ScanStatement statement;
+public class SortedBindableTableScanOp extends SimpleScanOp {
 
     public SortedBindableTableScanOp(ScanStatement scan) {
-        this.statement = scan;
-    }
-
-    @Override
-    public String getTablespace() {
-        return statement.getTableSpace();
-    }
-
-    public ScanStatement getStatement() {
-        return statement;
-    }
-
-    @Override
-    public StatementExecutionResult execute(TableSpaceManager tableSpaceManager,
-        TransactionContext transactionContext,
-        StatementEvaluationContext context, boolean lockRequired, boolean forWrite) throws StatementExecutionException {
-        DataScanner scan = tableSpaceManager.scan(statement, context, transactionContext, lockRequired, forWrite);
-        return new ScanResult(transactionContext.transactionId, scan);
-    }
-
-    @Override
-    public <T> T unwrap(Class<T> clazz) {
-        T unwrapped = statement.unwrap(clazz);
-        if (unwrapped != null) {
-            return unwrapped;
-        }
-        return Wrapper.unwrap(this, clazz);
+        super(scan);
     }
 
     @Override
